+++ examples/Kemper Profiler Player/MIDICaptain Mini 6/kitarist/display.py
@@ -89,8 +89,7 @@ Display = ParameterCondition(
         
         scale = 3,
         layout = {
-            "font": "/fonts/PTSans-NarrowBold-40.pcf",
-            "text": "Tuner"
+            "font": "/fonts/PTSans-NarrowBold-40.pcf"
         }
     )
 )