+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/FX States and Amp Name/display.py
@@ -100,8 +100,7 @@ Display = ParameterCondition(
         
         scale = 3,
         layout = {
-            "font": "/fonts/PTSans-NarrowBold-40.pcf",
-            "text": "Tuner"
+            "font": "/fonts/PTSans-NarrowBold-40.pcf"
         }
     )
 )