+++ test/pyswitch/test_memory.py
@@ -13,6 +13,7 @@ with patch.dict(sys.modules, {
     "gc": MockGC()
 }):
     from .mocks_misc import MockMisc
+    from gc import gc_mock_data
 
     with patch.dict(sys.modules, {
         "lib.pyswitch.misc": MockMisc
@@ -25,30 +26,31 @@ class TestMemory(unittest.TestCase):
 
     def test_watch(self):
         MockMisc.reset_mock()
+        gc_mock_data().reset()
 
-        MockGC.mock["memFreeReturn"] = 1000
-        MockGC.mock["memAllocReturn"] = 3333
+        gc_mock_data().output_mem_free = 1024 * 1024
+        gc_mock_data().output_mem_alloc = 3333
         
         Memory.start("foo")
         
         self.assertEqual(len(MockMisc.msgs), 1)
         self.assertIn("foo", MockMisc.latest_msg())     
-        self.assertIn(MockMisc.format_size(1000), MockMisc.latest_msg())        # Free memory
-        self.assertIn(MockMisc.format_size(1000 + 3333), MockMisc.latest_msg()) # Total memory
+        self.assertIn(MockMisc.format_size(1024 * 1024), MockMisc.latest_msg())        # Free memory
+        self.assertIn(MockMisc.format_size(1024 * 1024 + 3333), MockMisc.latest_msg()) # Total memory
 
         # Use 333 bytes
-        MockGC.mock["memFreeReturn"] = 777
+        gc_mock_data().output_mem_free = 1024 * 1024 - 333
         
         Memory.watch("bar")
         
         self.assertEqual(len(MockMisc.msgs), 2)
         self.assertIn("bar", MockMisc.latest_msg())     
         self.assertIn("Allocated", MockMisc.latest_msg())     
-        self.assertIn(MockMisc.format_size(777), MockMisc.latest_msg())            # Free memory
-        self.assertIn(MockMisc.format_size(1000 - 777), MockMisc.latest_msg())     # Total memory
+        self.assertIn(MockMisc.format_size(1024 * 1024 - 333), MockMisc.latest_msg())            # Free memory
+        self.assertIn(MockMisc.format_size(333), MockMisc.latest_msg())     # Total memory
 
         # Use 0 bytes
-        MockGC.mock["memFreeReturn"] = 777
+        gc_mock_data().output_mem_free = 1024 * 1024 - 333
         
         Memory.watch("bar2")
         
@@ -56,17 +58,17 @@ class TestMemory(unittest.TestCase):
         self.assertIn("bar2", MockMisc.latest_msg())     
         self.assertNotIn("Allocated", MockMisc.latest_msg())     
         self.assertNotIn("Released", MockMisc.latest_msg())     
-        self.assertIn(MockMisc.format_size(777), MockMisc.latest_msg())            # Free memory
-        self.assertNotIn(MockMisc.format_size(1000 - 777), MockMisc.latest_msg())  # Total memory
+        self.assertIn(MockMisc.format_size(1024 * 1024 - 333), MockMisc.latest_msg())            # Free memory
+        self.assertNotIn(MockMisc.format_size(333), MockMisc.latest_msg())     # Total memory
 
         # Free 100 bytes
-        MockGC.mock["memFreeReturn"] = 877
+        gc_mock_data().output_mem_free = 1024 * 1024 - 233
         
         Memory.watch("bar3")
         
         self.assertEqual(len(MockMisc.msgs), 4)
         self.assertIn("bar3", MockMisc.latest_msg())     
         self.assertIn("Released", MockMisc.latest_msg())     
-        self.assertIn(MockMisc.format_size(877), MockMisc.latest_msg())            # Free memory
-        self.assertNotIn(MockMisc.format_size(1000 - 877), MockMisc.latest_msg())  # Total memory
+        self.assertIn(MockMisc.format_size(1024 * 1024 - 233), MockMisc.latest_msg())            # Free memory
+        self.assertIn(MockMisc.format_size(100), MockMisc.latest_msg())  # Total memory